+++ release/build_release.sh
@@ -16,6 +16,8 @@ if [ -z "$RELEASE_BRANCH" ]; then
   exit 1
 fi
 
+BUILD_BRANCH=release-mici-staging
+
 
 # set git identity
 source $DIR/identity.sh
@@ -26,7 +28,7 @@ mkdir -p $BUILD_DIR
 cd $BUILD_DIR
 git init
 git remote add origin git@github.com:commaai/openpilot.git
-git checkout --orphan $RELEASE_BRANCH
+git checkout --orphan $BUILD_BRANCH
 
 # do the files copy
 echo "[-] copying files T=$SECONDS"
@@ -46,14 +48,14 @@ git commit -a -m "openpilot v$VERSION release"
 
 # Build
 export PYTHONPATH="$BUILD_DIR"
-scons -j$(nproc) --minimal
+scons
 
 if [ -z "$PANDA_DEBUG_BUILD" ]; then
   # release panda fw
-  CERT=/data/pandaextra/certs/release RELEASE=1 scons -j$(nproc) panda/
+  CERT=/data/pandaextra/certs/release RELEASE=1 scons panda/
 else
   # build with ALLOW_DEBUG=1 to enable features like experimental longitudinal
-  scons -j$(nproc) panda/
+  scons panda/
 fi
 
 # Ensure no submodules in release
@@ -72,14 +74,7 @@ find . -name '*.pyc' -delete
 find . -name 'moc_*' -delete
 find . -name '__pycache__' -delete
 rm -rf .sconsign.dblite Jenkinsfile release/
-rm -f selfdrive/modeld/models/*.onnx
-
-find third_party/ -name '*x86*' -exec rm -r {} +
-find third_party/ -name '*Darwin*' -exec rm -r {} +
-
-
-# Restore third_party
-git checkout third_party/
+rm -f selfdrive/modeld/models/*.onnx*
 
 # Mark as prebuilt release
 touch prebuilt
@@ -93,9 +88,11 @@ cd $BUILD_DIR
 RELEASE=1 pytest -n0 -s selfdrive/test/test_onroad.py
 #pytest selfdrive/car/tests/test_car_interfaces.py
 
-if [ ! -z "$RELEASE_BRANCH" ]; then
-  echo "[-] pushing release T=$SECONDS"
-  git push -f origin $RELEASE_BRANCH:$RELEASE_BRANCH
-fi
+echo "[-] pushing release T=$SECONDS"
+REFS=()
+for branch in ${RELEASE_BRANCH//,/ }; do
+  REFS+=("$BUILD_BRANCH:$branch")
+done
+git push -f origin "${REFS[@]}"
 
 echo "[-] done T=$SECONDS"