+++ system/athena/registration.py
@@ -95,7 +95,7 @@ def register(show_spinner=False) -> str | None:
       spinner.close()
 
   if dongle_id:
-    params.put("DongleId", dongle_id)
+    params.put("DongleId", dongle_id, block=True)
     set_offroad_alert("Offroad_UnregisteredHardware", (dongle_id == UNREGISTERED_DONGLE_ID) and not PC)
   return dongle_id