+++ system/manager/process.py
@@ -190,12 +190,8 @@ class PythonProcess(ManagerProcess):
     if self.proc is not None:
       return
 
-    # TODO: this is just a workaround for this tinygrad check:
-    # https://github.com/tinygrad/tinygrad/blob/ac9c96dae1656dc220ee4acc39cef4dd449aa850/tinygrad/device.py#L26
-    name = self.name if "modeld" not in self.name else "MainProcess"
-
     cloudlog.info(f"starting python {self.module}")
-    self.proc = Process(name=name, target=self.launcher, args=(self.module, self.name))
+    self.proc = Process(name=self.name, target=self.launcher, args=(self.module, self.name))
     self.proc.start()
     self.shutting_down = False
 
@@ -240,7 +236,7 @@ class DaemonProcess(ManagerProcess):
                                stderr=open('/dev/null', 'w'),
                                preexec_fn=os.setpgrp)
 
-    self.params.put(self.param_name, proc.pid)
+    self.params.put(self.param_name, proc.pid, block=True)
 
   def stop(self, retry=True, block=True, sig=None) -> None:
     pass