+++ RELEASES.md
@@ -1,7 +1,13 @@
-Version 0.11.1 (2026-04-22)
+Version 0.11.2 (2026-06-15)
+========================
+
+
+Version 0.11.1 (2026-05-18)
 ========================
 * New driver monitoring model
 * Improved image processing pipeline for driver camera
+* Improved thermal policy for comma four
+* Acura MDX 2022-24 support thanks to mvl-boston!
 * Rivian R1S and R1T 2025 support thanks to lukasloetkolben!
 
 Version 0.11.0 (2026-03-17)